refactor(cli): standardise the parameter names of all the cli tools
